# chacha20: add SSE2 accelerated variant