@tarcieri 2019-10-18
> SSE2 variant shows a good improvement over the scalar code, speeding up the benchmarks from 260MB/s to 327MB/s.

Awesome, this looks good at first glance.

> I'd also be keen to help adding support for AVX2 and AVX512, but it looks like this may need larger changes (possibly in the salsa20_core crate) to allow for processing multiple blocks simultaneously?

Yes. It might make sense to look at ways to improve/simplify that crate as well, as its implementation is rather complex and not particularly optimal. Perhaps it would benefit from taking a 32-bit counter and outputting 512-bit blocks as `[u8; 64]`, rather than the current approach which operates directly on the `[u32; 8]` array.

Additionally note we haven't implemented an optimization where the ChaCha (or Salsa) where the input array / block state is computed once, and then copied with only the counter updated to avoid setting it up again over and over for each block, as described in [RFC 8439 Section 3](https://tools.ietf.org/html/rfc8439#section-3)

> Each block of ChaCha20 involves 16 move operations and one increment operation for loading the state, 80 each of XOR, addition and roll operations for the rounds, 16 more add operations and 16 XOR operations for protecting the plaintext.  Section 2.3 describes the ChaCha block function as "adding the original input words".  This implies that before starting the rounds on the ChaCha state, we copy it aside, only to add it in later.  This is correct, but we can save a few operations if we instead copy the state and do the work on the copy.  This way, for the next block you don't need to recreate the state, but only to increment the block counter.  This saves approximately 5.5% of the cycles.

AVX2 sounds good.  As for AVX-512, [I had suspected that in most situations the performance tradeoffs for AVX-512 aren't worth it](https://moderncrypto.org/mail-archive/noise/2017/001208.html) only to have that [later confirmed by Cloudflare engineers](https://blog.cloudflare.com/on-the-dangers-of-intels-frequency-scaling/). An AVX-512 implementation of ChaCha20 is really only useful as something like a CSPRNG feeding another SIMD workload. As soon as you try to combine it into the ChaCha20Poly1305 AEAD, performance suffers, so I'd recommend just stopping with AVX2.

Let me get back to your questions on gating / facade / testing. I had lots of fun with these in the [`polyval` crate](https://github.com/RustCrypto/universal-hashes/tree/master/polyval), but don't have time to write them up right now.