@tarcieri 2019-10-18
Answering your remaining questions...

### Gating

> I've only implemented compile-time feature detection for now, as that seemed easier. I could potentially add runtime detection as well, if you think that is useful?

I'd suggest getting everything working with compile-time gating only and then circling back on runtime gating. [I ended up abandoning runtime feature detection for PCLMULQDQ in the `polyval` crate](https://github.com/RustCrypto/universal-hashes/pull/11), but may end up circling back on it eventually.

> Additionally, I was considering to pull in cfg-if to simplify the detection logic, but wasn't sure if you wanted to have an additional dependency. Lastly, do we need to be backwards compatible to Rust versions before core::arch got stabilized?

I'm fine with pulling in `cfg-if` if you find it helpful. However, I tried to use it to simplify the gating in the `polyval` PR linked above only to discover it was too inexpressive to really help simplify the gating there.

### Facade

> I'm currently switching between the two implementations inside the cipher module, but I wonder if there is a better way? Potentially, the block module could become a facade for block/scalar and block/sse2, which would abstract away the fact that there are multiple implementations?

In `polyval` there's a `field::Element` facade/newtype, but I've been wondering if that might make sense to replace with a trait or thereabouts.

I think you could potentially have an `SseCipher` type which impls the same API, which the `ChaCha20` type consumes directly.

### Testing

> I've opted to pregenerate random inputs that I use in the tests, which seemed like an okay approach. I was considering pulling in proptest to essentially randomly test the new variant against the scalar implementation, but I wasn't quite sure if you wanted to have it as a dependency.

Add whatever `dev-dependencies` you like. That sounds interesting.

That said, I originally did something similar in the `polyval` crate only to abandon it after changing the conditional gating so only one implementation ends up getting compiled into the crate. I tried to support multiple implementations in the runtime feature detection PR, but abandoned it.

tl;dr: I've gone back and forth on a lot of this stuff, and don't have good answers or strong preferences either way.