@srijs 2019-10-19
Thanks for the feedback! After trying a few different approaches, I've now settled on the following choices:

- Moved the SSE2 implementation to be a submodule of the `block` module.
- Added a comparison test to check the SIMD implementation against the scalar one.
- Kept the feature detection static, because making the `#[no_std]` directive optional via feature flag is probably a bit more work.
- Kept the feature branching inside `cipher`, because the module seemed simple enough and I'm not quite sure this really needs more facade.
- Did not pull in `cfg-if`, because it is incompatible with Rust 1.27 (CI was failing).
- Did not pull in `proptest`, since it needs `std` to work properly (or at least `alloc` on nightly).

I think I'm quite happy with the changes for now, and would defer to you for any change requests you may have :)