@tarcieri 2019-10-19
Looks good to me and I'll go ahead and merge. Some notes:

> Kept the feature detection static, because making the #[no_std] directive optional via feature flag is probably a bit more work.

Yeah, this was a big pain in the ass for my (abandoned) [`polyval` runtime feature detection for PCLMULQDQ](https://github.com/RustCrypto/universal-hashes/pull/11) PR as well. I had the `std` feature off-by-default before that.

> Did not pull in cfg-if, because it is incompatible with Rust 1.27 (CI was failing).

Oof, yeah at some point I think we can look into doing a 2018 edition upgrade for select crates. It was easier to keep them consistent at first, but it's certainly getting annoying to deal with such old Rust versions.