@tarcieri 2019-10-19
@srijs do you want to do an AVX2 implementation before I do another release?

I'm going to throw `criterion` with the cycles-per-byte plugin on this to get some better benchmarks.