@tarcieri 2019-10-19
Also, looking at the optimization described in [RFC8439 Section 3](https://tools.ietf.org/html/rfc8439#section-3), it'd probably be good to do that before adding any additional backends, since that impacts how they're structured.